Add simple argument modification but will change soon
In current implementation, each step is a user (instruction or argument).
This is not suitable for searching data flow (dst) and finding definition (src)
at the same time because we lose the last instruction information when
processing StoreInst, and thus we do not need whether it was used as the
src operand or dst operand.

In the immediate next implementation, we will change the implementation
of processUser: it will process users in the users loop, instead of
process it at outside. The new processUser will have two parts: 1) search
for dst when as src operand in the user, 2) search for src when as dst operand.
I expect there should have no StoreInst for each step. Additionally, we
will add field offset chain.